+++ FE/apis/roleApi.ts
@@ -28,21 +28,21 @@ const request = async (path: string, options: RequestInit = {}): Promise<any> =>
 
 // 역할 리스트 조회
 export const getRoles = async (workspaceId: string) => {
-  return request(`http://${BASE}/api/workspaces/${workspaceId}/roles`, {
+  return request(`${BASE}/api/workspaces/${workspaceId}/roles`, {
     method: "GET",
   });
 };
 
 // 역할별 권한 조회
 export const getRoleById = async (workspaceId: string, roleId: string) => {
-  return request(`http://${BASE}/api/workspaces/${workspaceId}/roles/${roleId}`, {
+  return request(`${BASE}/api/workspaces/${workspaceId}/roles/${roleId}`, {
     method: "GET",
   });
 };
 
 // 역할별 권한 생성
 export const createRole = async (workspaceId: string, roleName: string, rolePermission: any) => {
-  return request(`http://${BASE}/api/workspaces/${workspaceId}/roles`, {
+  return request(`${BASE}/api/workspaces/${workspaceId}/roles`, {
     method: "POST",
     body: JSON.stringify({ role_name: roleName, role_permission: rolePermission }),
   });
@@ -50,7 +50,7 @@ export const createRole = async (workspaceId: string, roleName: string, rolePerm
 
 // 역할별 권한 수정
 export const updateRole = async (workspaceId: string, roleName: string, rolePermission: any) => {
-  return request(`http://${BASE}/api/workspaces/${workspaceId}/roles`, {
+  return request(`${BASE}/api/workspaces/${workspaceId}/roles`, {
     method: "PATCH",
     body: JSON.stringify({ role_name: roleName, role_permission: rolePermission }),
   });
@@ -58,7 +58,7 @@ export const updateRole = async (workspaceId: string, roleName: string, rolePerm
 
 // 역할 삭제
 export const deleteRole = async (workspaceId: string, roleName: string) => {
-  return request(`http://${BASE}/api/workspaces/${workspaceId}/roles/delete`, {
+  return request(`${BASE}/api/workspaces/${workspaceId}/roles/delete`, {
     method: "PATCH",
     body: JSON.stringify({ role_name: roleName }),
   });
